main/requirements.txt
weaviate-client[agents]>=4.20.0
python-dotenv>=1.0.1
Flask>=3.0.0
